Query: Move to different submodule
and integrate Query into the new module structure. It is now also
implemented using dispatch and the new API functions `to_string` and
friends. The old-style API has been kept intact.
