+++ src/De/De.jl
@@ -88,6 +88,3 @@ to_deser(::Type{T}, x) where {T} = deser(deser_type(T, x), deser_value(T, x))
 
 to_deser(::Type{Nothing}, x) = nothing
 to_deser(::Type{Missing}, x) = missing
-
-include("DeQuery.jl")
-using .DeQuery
