+++ src/Par/Par.jl
@@ -1,4 +1,2 @@
 # Par/Par
 
-include("ParQuery.jl")
-using .ParQuery
