+++ src/Ser/Ser.jl
@@ -6,6 +6,3 @@
 
 (ser_ignore_field(::Type{T}, ::Val{x})::Bool) where {T,x} = false
 (ser_ignore_field(::Type{T}, k::Val{x}, v::V)::Bool) where {T,x,V} = ser_ignore_field(T, k)
-
-include("SerQuery.jl")
-using .SerQuery
